+++ tosho_kmkc/src/lib.rs
@@ -184,6 +184,8 @@ impl KMClient {
 
                 // make cookie store
                 let client = reqwest::Client::builder()
+                    .http2_adaptive_window(true)
+                    .use_rustls_tls()
                     .default_headers(headers)
                     .cookie_provider(std::sync::Arc::clone(&cookie_store));
 
@@ -210,6 +212,8 @@ impl KMClient {
                 let cookie_store = std::sync::Arc::new(cookie_store);
 
                 let client = reqwest::Client::builder()
+                    .http2_adaptive_window(true)
+                    .use_rustls_tls()
                     .default_headers(headers)
                     .cookie_provider(std::sync::Arc::clone(&cookie_store));
 
@@ -879,6 +883,8 @@ impl KMClient {
         let cookie_store = std::sync::Arc::new(cookie_store);
 
         let client = reqwest::Client::builder()
+            .http2_adaptive_window(true)
+            .use_rustls_tls()
             .default_headers(headers)
             .cookie_provider(std::sync::Arc::clone(&cookie_store))
             .build()?;