+++ tosho_musq/src/lib.rs
@@ -146,7 +146,10 @@ impl MUClient {
             reqwest::header::HeaderValue::from_str(&constants.api_ua).unwrap(),
         );
 
-        let client = reqwest::Client::builder().default_headers(headers);
+        let client = reqwest::Client::builder()
+            .http2_adaptive_window(true)
+            .use_rustls_tls()
+            .default_headers(headers);
 
         let client = match proxy {
             Some(proxy) => client.proxy(proxy).build().unwrap(),