+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Starting from Rust port of the project, all changes will be put into this file.
 ## Unreleased (git master)
 ### Changes
 - All source: Force use `rustls` and use `http2` adaptive window for reqwest client.
+- `MU`: Update proto, move `Unrecognized` enums to the top of the list as default.
 
 ### Build
 - Bump `reqwest` to 0.12