+++ CHANGELOG.md
@@ -11,12 +11,14 @@ Starting from Rust port of the project, all changes will be put into this file.
 - `MU`: Update proto information.
 - `MU`: Only use `HTTP/1.1` since `HTTP/2` broke the API request
 - `MU`: Rework downloader, image blocks are now stored to save request
+- `RB`: Fix wrong base host used in homepage view
 - Refactor some duplicate code
 
 ### Build
 - Bump `reqwest` to 0.12
 - Use `rustls` instead of `native-tls` for `reqwest`
 - Set `stream` feature as default on `reqwest`
+- Compile with latest Rust stable in CI
 - Bump other dependencies
 
 ## [0.4.3] (2024-03-20)